requirements.txt
# Bot dependencies
python-telegram-bot~=13.15
python-dotenv>=1.0.0
requests>=2.31.0

# Database
tinydb>=4.8.0

# Data visualization
plotext>=5.2.8  # Lightweight plotting library
